hackon/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-dotenv
orjson
numpy
pandas
rapidfuzz
scikit-learn
networkx

pypdf
PyMuPDF
beautifulsoup4
lxml
markdown

pillow
pytesseract
pdf2image

sentence-transformers
FlagEmbedding

openai
